Manually add in versions of the timeago info for our 'fake' languages.
Summary:
I decided this was easier than marking up a template file with
$._("...") and letting fake_translate.py do the translation: I didn't
know how to handle the %d's in these format strings.  I may end up
regretting that if these files change a lot, but it will probably be
ok.

Manually add in versions of the timeago info for our 'fake' languages.

I decided this was easier than marking up a template file with
$._("...") and letting fake_translate.py do the translation: I didn't
know how to handle the %d's in these format strings.  I may end up
regretting that if these files change a lot, but it will probably be
ok.

Fix this file to make it actually utf-8.

Add a symlink to the 'correct' locale file for the current language.

This defaults to English.  When we copy this directory tree over to
other languages, we'll magically change this symlink to point to the
right file for that language.

Test Plan:
Fingers crossed.

Fingers crossed.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D2152