Move to __language__ rather than [language].
Firefox has problems fetching files with [ in their name -- it
url-encodes the filename component, even though it probably shouldn't,
and appengine can't handle that, even though it probably should.

In practice this is only a problem with dev-appserver, since in prod
this file is incorporated into a _hashed.js file, but better to avoid
the possibility of trouble later.

Auditor: john,chris